修正因解决上一版本错误（SHA-1: 42e65ac6475063799239ac8ef92b4461cc8d3c41）而引起的timer数据在跨服务器跳转时出错的问题，这个问题会导致两个cellapp的连接通道断开

+++ kbe/src/server/cellapp/witness.cpp
@@ -93,8 +93,8 @@ void Witness::addToStream(KBEngine::MemoryStream& s)
 	*/
 
 	// 当前这么做能解决问题，但是在space多cell分割的情况下将会出现问题
-	s << aoiRadius_ << aoiHysteresisArea_ << 0;	
-	s << 0; // aoiEntities_.size();
+	s << aoiRadius_ << aoiHysteresisArea_ << (uint16)0;	
+	s << (uint32)0; // aoiEntities_.size();
 }
 
 //-------------------------------------------------------------------------------------
